fix(ui): address PR review — breadcrumb nav, accessibility, file split
- Fix breadcrumb navigation: carry original path alongside formatted
  label so clicking truncated ID segments navigates correctly
- Add aria-expanded to sidebar group toggles
- Remove empty <p> spacer when sidebar is collapsed
- Add visibility:hidden to collapsed group items for keyboard a11y
- Extract SidebarFooter and breadcrumb logic into separate files
  to bring DashboardLayout.tsx under 300 lines (267)
