+++ packages/client/src/layouts/DashboardLayout.tsx
@@ -1,50 +1,16 @@
 import { useEffect, useState } from "react";
 import { NavLink, Navigate, Outlet, useLocation, useNavigate } from "react-router-dom";
 import { ErrorBoundary } from "../components/ErrorBoundary";
-import { useTheme } from "../components/theme-context";
-import { useSession, authClient } from "../lib/auth-client";
+import { useSession } from "../lib/auth-client";
 import { Separator } from "@/components/ui/separator";
-import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { Skeleton } from "@/components/ui/skeleton";
 import { KeyboardShortcutsDialog } from "@/components/KeyboardShortcutsDialog";
 import { CommandPalette } from "@/components/CommandPalette";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger
-} from "@/components/ui/dropdown-menu";
-import {
-  ChevronLeft,
-  ChevronRight,
-  ChevronDown,
-  LogOut,
-  User,
-  ChevronsUpDown,
-  Hexagon,
-  Sun,
-  Moon,
-  Menu
-} from "lucide-react";
+import { ChevronLeft, ChevronRight, ChevronDown, Hexagon, Menu } from "lucide-react";
 import { homeNavGroups, settingsNav } from "./sidebar-nav";
-
-const ID_SEGMENT_RE = /^[0-9a-f]{9,}$|^[0-9a-f-]{20,}$/i;
-
-function formatSegment(s: string): string {
-  if (ID_SEGMENT_RE.test(s)) return s.slice(0, 8) + "…";
-  return s
-    .split("-")
-    .map((w) => w.charAt(0).toUpperCase() + w.slice(1))
-    .join(" ");
-}
-
-function breadcrumbFromPath(pathname: string): string[] {
-  if (pathname === "/") return ["Dashboard"];
-  return pathname.split("/").filter(Boolean).map(formatSegment);
-}
+import { breadcrumbFromPath } from "./dashboard-breadcrumb";
+import { SidebarFooter } from "./SidebarFooter";
 
 export function DashboardLayout() {
   const session = useSession();
@@ -52,7 +18,6 @@ export function DashboardLayout() {
   const navigate = useNavigate();
   const [collapsed, setCollapsed] = useState(false);
   const [mobileOpen, setMobileOpen] = useState(false);
-  const themeCtx = useTheme();
   const crumbs = breadcrumbFromPath(location.pathname);
 
   const [openGroups, setOpenGroups] = useState<Record<string, boolean>>(() => {
@@ -114,8 +79,6 @@ export function DashboardLayout() {
     return <Navigate to={`/login?returnTo=${encodeURIComponent(returnTo)}`} replace />;
   }
 
-  const userInitial = session.data.user.name?.charAt(0).toUpperCase() ?? "U";
-
   function isSettingsItemActive(tab: string | null) {
     if (location.pathname !== "/settings") {
       return false;
@@ -167,11 +130,10 @@ export function DashboardLayout() {
                 const isOpen = openGroups[group.key] !== false;
                 return (
                   <div key={group.key}>
-                    {collapsed ? (
-                      <p className="sidebar__group-label" />
-                    ) : (
+                    {!collapsed && (
                       <button
                         className="sidebar__group-toggle"
+                        aria-expanded={isOpen}
                         data-open={isOpen ? "true" : "false"}
                         onClick={() => toggleGroup(group.key)}
                       >
@@ -241,79 +203,11 @@ export function DashboardLayout() {
             </nav>
           </TooltipProvider>
 
-          <TooltipProvider delay={0}>
-            <div className="sidebar__footer">
-              {(() => {
-                const btn = (
-                  <button
-                    className="sidebar__link"
-                    onClick={() => {
-                      const { resolved, setTheme } = themeCtx;
-                      setTheme(resolved === "dark" ? "light" : "dark");
-                    }}
-                  >
-                    {themeCtx.resolved === "dark" ? (
-                      <Sun size={18} className="sidebar__link-icon" />
-                    ) : (
-                      <Moon size={18} className="sidebar__link-icon" />
-                    )}
-                    {!collapsed && (
-                      <span>{themeCtx.resolved === "dark" ? "Light mode" : "Dark mode"}</span>
-                    )}
-                  </button>
-                );
-                return collapsed ? (
-                  <Tooltip>
-                    <TooltipTrigger render={btn} />
-                    <TooltipContent side="right">Toggle theme</TooltipContent>
-                  </Tooltip>
-                ) : (
-                  btn
-                );
-              })()}
-              <DropdownMenu>
-                <DropdownMenuTrigger
-                  render={
-                    <button className="sidebar__user-card group">
-                      <Avatar className="h-8 w-8 ring-2 ring-transparent transition-all group-hover:ring-primary/20">
-                        <AvatarFallback className="bg-primary/10 text-xs font-semibold text-primary">
-                          {userInitial}
-                        </AvatarFallback>
-                      </Avatar>
-                      {!collapsed && (
-                        <>
-                          <div className="sidebar__user-info">
-                            <p className="sidebar__user-name">{session.data.user.name}</p>
-                            <p className="sidebar__user-email">{session.data.user.email}</p>
-                          </div>
-                          <ChevronsUpDown size={14} className="ml-auto opacity-50" />
-                        </>
-                      )}
-                    </button>
-                  }
-                />
-                <DropdownMenuContent side="top" className="w-56 backdrop-blur-xl" align="start">
-                  <DropdownMenuLabel>
-                    <p className="font-medium">{session.data.user.name}</p>
-                    <p className="text-xs text-muted-foreground">{session.data.user.email}</p>
-                  </DropdownMenuLabel>
-                  <DropdownMenuSeparator />
-                  <DropdownMenuItem onClick={() => void navigate("/profile")}>
-                    <User size={14} />
-                    Profile Settings
-                  </DropdownMenuItem>
-                  <DropdownMenuSeparator />
-                  <DropdownMenuItem
-                    onClick={() => void authClient.signOut()}
-                    className="text-destructive"
-                  >
-                    <LogOut size={14} />
-                    Sign out
-                  </DropdownMenuItem>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </div>
-          </TooltipProvider>
+          <SidebarFooter
+            collapsed={collapsed}
+            userName={session.data.user.name ?? ""}
+            userEmail={session.data.user.email}
+          />
         </aside>
 
         {/* ── Main content ── */}
@@ -328,24 +222,18 @@ export function DashboardLayout() {
             </button>
             <nav className="topbar__breadcrumb" aria-label="Breadcrumb">
               {crumbs.map((crumb, i) => {
-                const path =
-                  "/" +
-                  crumbs
-                    .slice(0, i + 1)
-                    .map((c) => c.toLowerCase())
-                    .join("/");
                 const isLast = i === crumbs.length - 1;
                 return (
-                  <span key={crumb}>
+                  <span key={crumb.path}>
                     {i > 0 && <span className="topbar__breadcrumb-sep">/</span>}
                     {isLast ? (
-                      <span className="topbar__breadcrumb-current">{crumb}</span>
+                      <span className="topbar__breadcrumb-current">{crumb.label}</span>
                     ) : (
                       <button
                         className="topbar__breadcrumb-item hover:underline"
-                        onClick={() => void navigate(path === "/dashboard" ? "/" : path)}
+                        onClick={() => void navigate(crumb.path)}
                       >
-                        {crumb}
+                        {crumb.label}
                       </button>
                     )}
                   </span>
